requirements.txt
gradio==4.0.0
transformers>=4.40.0
torch>=2.0.0
safetensors>=0.4.0
accelerate>=0.27.0
datasets>=2.18.0
protobuf>=3.20.0
pyyaml>=6.0
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
huggingface_hub>=0.19.0